Add threading to writer, output.
Disabled writing values by default.

Updated examples.

Improved output display.

Improved debugging output.

Line formatting, etc.

Disabled outputting Gyro info in example, still displays headset info though including that information. With the old code commented. Printing to the console takes a fair amount of processor time, best not to do it twice.

Renamed imports to help with development if you have emokit.whatever and the emokit package is installed when you run it, it will look at the package version instead of the dev version so we use the . instead.

Added coding statement to various files.

Hopefully fixed some things!

Added output_path option but, have not tested it. Hopefully it works.

Incremented version number.

Moved EmotivWriter out of .util to it's own file.

Created Tasks to pass into the various Queues so we can add extra information, like when the data was received.

Python3 exporting is not nearly as stable as Python2 I believe this is a language thing.